Implement FELighting applier for Neon https://bugs.webkit.org/show_bug.cgi?id=267349 rdar://120791073
Reviewed by Simon Fraser.

Add a virtual method to FELightingSoftwareApplier named applyPlatformParallel() and
call it from its applyPlatform(). Add two new classes FELightingSoftwareParallelApplier
and FELightingNeonParallelApplier and make each one of them implement this method. Move
the existing code in FELightingNEON to the new class FELightingNeonParallelApplier.
Make FELightingSoftwareParallelApplier for all other platforms.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/cpu/arm/filters/FELightingNEON.h: Removed.
* Source/WebCore/platform/graphics/cpu/arm/filters/FELightingNeonParallelApplier.cpp: Renamed from Source/WebCore/platform/graphics/cpu/arm/filters/FELightingNEON.cpp.
(WebCore::FELightingNeonParallelApplier::feLightingConstants):
(WebCore::FELightingNeonParallelApplier::getPowerCoefficients):
(WebCore::FELightingNeonParallelApplier::applyPlatformWorker):
(WebCore::FELightingNeonParallelApplier::applyPlatformParallel const):
* Source/WebCore/platform/graphics/cpu/arm/filters/FELightingNeonParallelApplier.h: Added.
* Source/WebCore/platform/graphics/filters/FELighting.cpp:
(WebCore::FELighting::createSoftwareApplier const):
* Source/WebCore/platform/graphics/filters/software/FECompositeSoftwareArithmeticApplier.cpp:
* Source/WebCore/platform/graphics/filters/software/FELightingSoftwareApplier.cpp:
(WebCore::FELightingSoftwareApplier::applyPlatform const):
(WebCore::FELightingSoftwareApplier::applyPlatformGenericPaint): Deleted.
(WebCore::FELightingSoftwareApplier::applyPlatformGenericWorker): Deleted.
(WebCore::FELightingSoftwareApplier::applyPlatformGeneric): Deleted.
(WebCore::FELightingSoftwareApplier::applyPlatform): Deleted.
* Source/WebCore/platform/graphics/filters/software/FELightingSoftwareApplier.h:
* Source/WebCore/platform/graphics/filters/software/FELightingSoftwareParallelApplier.cpp: Added.
(WebCore::FELightingSoftwareParallelApplier::applyPlatformPaint):
(WebCore::FELightingSoftwareParallelApplier::applyPlatformWorker):
(WebCore::FELightingSoftwareParallelApplier::applyPlatformParallel const):
* Source/WebCore/platform/graphics/filters/software/FELightingSoftwareParallelApplier.h: Added.

Canonical link: https://commits.webkit.org/272873@main
