+++ Source/WebCore/platform/graphics/filters/software/FELightingSoftwareApplier.cpp
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) 2010 University of Szeged
  * Copyright (C) 2010 Zoltan Herczeg
- * Copyright (C) 2018-2022 Apple, Inc. All rights reserved.
+ * Copyright (C) 2018-2024 Apple, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -32,7 +32,6 @@
 #include "Filter.h"
 #include "ImageBuffer.h"
 #include "PixelBuffer.h"
-#include <wtf/ParallelJobs.h>
 
 namespace WebCore {
 
@@ -245,85 +244,7 @@ void FELightingSoftwareApplier::setPixel(int offset, const LightingData& data, c
     setPixelInternal(offset, data, paintingData, x, y, factorX, factorY, normal2DVector, data.pixels->item(offset + cAlphaChannelOffset));
 }
 
-// This appears to read from and write to the same pixel buffer, but it only reads the alpha channel, and writes the non-alpha channels.
-void FELightingSoftwareApplier::applyPlatformGenericPaint(const LightingData& data, const LightSource::PaintingData& paintingData, int startY, int endY)
-{
-    // Make sure startY is > 0 since we read from the previous row in the loop.
-    ASSERT(startY);
-    ASSERT(endY > startY);
-
-    for (int y = startY; y < endY; ++y) {
-        int rowStartOffset = y * data.widthMultipliedByPixelSize;
-        int previousRowStart = rowStartOffset - data.widthMultipliedByPixelSize;
-        int nextRowStart = rowStartOffset + data.widthMultipliedByPixelSize;
-
-        // alphaWindow is a local cache of alpha values.
-        // Fill the two right columns putting the left edge value in the center column.
-        // For each pixel, we shift each row left then fill the right column.
-        AlphaWindow alphaWindow;
-        alphaWindow.setTop(data.pixels->item(previousRowStart + cAlphaChannelOffset));
-        alphaWindow.setTopRight(data.pixels->item(previousRowStart + cPixelSize + cAlphaChannelOffset));
-
-        alphaWindow.setCenter(data.pixels->item(rowStartOffset + cAlphaChannelOffset));
-        alphaWindow.setRight(data.pixels->item(rowStartOffset + cPixelSize + cAlphaChannelOffset));
-
-        alphaWindow.setBottom(data.pixels->item(nextRowStart + cAlphaChannelOffset));
-        alphaWindow.setBottomRight(data.pixels->item(nextRowStart + cPixelSize + cAlphaChannelOffset));
-
-        int offset = rowStartOffset + cPixelSize;
-        for (int x = 1; x < data.width - 1; ++x, offset += cPixelSize) {
-            alphaWindow.shift();
-            setPixelInternal(offset, data, paintingData, x, y, cFactor1div4, cFactor1div4, data.interiorNormal(offset, alphaWindow), alphaWindow.center());
-        }
-    }
-}
-
-void FELightingSoftwareApplier::applyPlatformGenericWorker(ApplyParameters* parameters)
-{
-    applyPlatformGenericPaint(parameters->data, parameters->paintingData, parameters->yStart, parameters->yEnd);
-}
-
-#if !(CPU(ARM_NEON) && CPU(ARM_TRADITIONAL) && COMPILER(GCC_COMPATIBLE))
-void FELightingSoftwareApplier::applyPlatformGeneric(const LightingData& data, const LightSource::PaintingData& paintingData)
-{
-    unsigned rowsToProcess = data.height - 2;
-    unsigned maxNumThreads = rowsToProcess / 8;
-    
-    static constexpr int minimalRectDimension = 100 * 100; // Empirical data limit for parallel jobs
-    unsigned optimalThreadNumber = std::min<unsigned>(((data.width - 2) * rowsToProcess) / minimalRectDimension, maxNumThreads);
-
-    if (optimalThreadNumber > 1) {
-        // Initialize parallel jobs
-        ParallelJobs<ApplyParameters> parallelJobs(&applyPlatformGenericWorker, optimalThreadNumber);
-
-        // Fill the parameter array
-        int job = parallelJobs.numberOfJobs();
-        if (job > 1) {
-            // Split the job into "yStep"-sized jobs but there a few jobs that need to be slightly larger since
-            // yStep * jobs < total size. These extras are handled by the remainder "jobsWithExtra".
-            const int yStep = rowsToProcess / job;
-            const int jobsWithExtra = rowsToProcess % job;
-
-            int yStart = 1;
-            for (--job; job >= 0; --job) {
-                ApplyParameters& params = parallelJobs.parameter(job);
-                params.data = data;
-                params.paintingData = paintingData;
-                params.yStart = yStart;
-                yStart += job < jobsWithExtra ? yStep + 1 : yStep;
-                params.yEnd = yStart;
-            }
-            parallelJobs.execute();
-            return;
-        }
-        // Fallback to single threaded mode.
-    }
-
-    applyPlatformGenericPaint(data, paintingData, 1, data.height - 1);
-}
-#endif
-
-void FELightingSoftwareApplier::applyPlatform(const LightingData& data)
+void FELightingSoftwareApplier::applyPlatform(const LightingData& data) const
 {
     LightSource::PaintingData paintingData;
 
@@ -374,7 +295,7 @@ void FELightingSoftwareApplier::applyPlatform(const LightingData& data)
 
     if (data.width >= 3 && data.height >= 3) {
         // Interior pixels.
-        applyPlatformGeneric(data, paintingData);
+        applyPlatformParallel(data, paintingData);
     }
 
     int lastPixel = data.widthMultipliedByPixelSize * data.height;
