+++ Source/WebCore/platform/graphics/filters/software/FELightingSoftwareApplier.h
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) 2010 University of Szeged
  * Copyright (C) 2010 Zoltan Herczeg
- * Copyright (C) 2018-2022 Apple, Inc. All rights reserved.
+ * Copyright (C) 2018-2024 Apple, Inc. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -37,16 +37,14 @@ namespace WebCore {
 
 class FELighting;
 
-class FELightingSoftwareApplier final : public FilterEffectConcreteApplier<FELighting> {
+class FELightingSoftwareApplier : public FilterEffectConcreteApplier<FELighting> {
     WTF_MAKE_FAST_ALLOCATED;
     using Base = FilterEffectConcreteApplier<FELighting>;
 
-public:
+protected:
     using Base::Base;
 
-private:
-    bool apply(const Filter&, const FilterImageVector& inputs, FilterImage& result) const final;
-
+    static constexpr int minimalRectDimension = 100 * 100; // Empirical data limit for parallel jobs
     static constexpr int cPixelSize = 4;
     static constexpr int cAlphaChannelOffset = 3;
     static constexpr uint8_t cOpaqueAlpha = static_cast<uint8_t>(0xFF);
@@ -120,20 +118,12 @@ private:
         inline IntSize bottomRightNormal(int offset) const;
     };
 
-    struct ApplyParameters {
-        LightingData data;
-        LightSource::PaintingData paintingData;
-        int yStart;
-        int yEnd;
-    };
-
     static void setPixelInternal(int offset, const LightingData&, const LightSource::PaintingData&, int x, int y, float factorX, float factorY, IntSize normal2DVector, float alpha);
     static void setPixel(int offset, const LightingData&, const LightSource::PaintingData&, int x, int y, float factorX, float factorY, IntSize normal2DVector);
 
-    static void applyPlatformGenericPaint(const LightingData&, const LightSource::PaintingData&, int startY, int endY);
-    static void applyPlatformGenericWorker(ApplyParameters*);
-    static void applyPlatformGeneric(const LightingData&, const LightSource::PaintingData&);
-    static void applyPlatform(const LightingData&);
+    virtual void applyPlatformParallel(const LightingData&, const LightSource::PaintingData&) const = 0;
+    void applyPlatform(const LightingData&) const;
+    bool apply(const Filter&, const FilterImageVector& inputs, FilterImage& result) const final;
 };
 
 } // namespace WebCore
